+++ leetcodePython/matrix/matrix_200.py
@@ -96,46 +96,13 @@ class SolutionDFSNoModify:
         return islands
 
 
-class UnionFind:
-    def __init__(self, grid: List[List[str]]):
-        self._count = 0
-        self._parent = []
-        self._rank = []
-        for row in range(len(grid)):
-            for col in range(len(grid[0])):
-                if grid[row][col] == "1":
-                    cell_index = row * len(grid[0]) + col
-                    self._parent.append(cell_index)
-                    self._count += 1
-                else:
-                    self._parent.append(-1)
-                self._rank.append(0)
-
-    def find(self, data: int) -> int:
-        if self._parent[data] != data:
-            self._parent[data] = self.find(self._parent[data])
-        return self._parent[data]
-
-    def union(self, a: int, b: int) -> None:
-        root_a = self.find(a)
-        root_b = self.find(b)
-        if root_a != root_b:
-            if self._rank[root_a] > self._rank[root_b]:
-                self._parent[root_b] = root_a
-            elif self._rank[root_a] < self._rank[root_b]:
-                self._parent[root_a] = root_b
-            else:
-                self._parent[root_b] = root_a
-                self._rank[root_a] += 1
-            self._count -= 1
-
-    def count(self) -> int:
-        return self._count
+from UnionFind.DataStructure import UnionFindInt
 
 
 class SolutionUnionFind:
     @staticmethod
-    def _join_new_land(grid: List[List[str]], uf: UnionFind, land: tuple[int, int], new_land: tuple[int, int]) -> None:
+    def _join_new_land(grid: List[List[str]], uf: UnionFindInt, land: tuple[int, int],
+                       new_land: tuple[int, int]) -> None:
         r_new, c_new = new_land
         if r_new < 0 or c_new < 0:
             return
@@ -153,7 +120,7 @@ class SolutionUnionFind:
             return 0
         rows = len(grid)
         cols = len(grid[0])
-        uf = UnionFind(grid)
+        uf = UnionFindInt(grid)
 
         for row in range(rows):
             for col in range(cols):
@@ -164,3 +131,10 @@ class SolutionUnionFind:
                 self._join_new_land(grid, uf, (row, col), (row, col - 1))
                 self._join_new_land(grid, uf, (row, col), (row, col + 1))
         return uf.count()
+
+
+if __name__ == "__main__":
+    sol = SolutionUnionFind()
+    ans = sol.numIslands(
+        [["1", "1", "1", "1", "0"], ["1", "1", "0", "1", "0"], ["1", "1", "0", "0", "0"], ["0", "0", "0", "0", "0"]])
+    print(ans)
